move _RetryQueue class to transport/utils.py
