+++ raiden/tests/integration/network/transport/test_matrix_transport.py
@@ -23,7 +23,8 @@ from raiden.utils.signer import LocalSigner
 from raiden.utils.typing import Address, List, Optional, Union
 from transport.matrix.client import Room
 from transport.matrix.node import MatrixNode as MatrixTransportNode
-from transport.matrix.utils import AddressReachability, make_room_alias, _RetryQueue
+from transport.matrix.utils import AddressReachability, make_room_alias
+from transport.utils import _RetryQueue
 from transport.message import Message as TransportMessage
 
 USERID0 = "@Arthur:RestaurantAtTheEndOfTheUniverse"
